Refactor file handling
The global variable "_fileToOpen" was replaced by a variable inside the
"windowManagement" module for the last opend file. "windowManagement"
is the only place that should know about which file to open in which
way under which circumstances. If no file is given to
"windowManagement.open", it will open the last opened file or if there
is none yet, the default file (README.md).