First approach for handling links
Behavior is changed to the old app: clicking on a link to an external
document always opens a new window while cliking on an internal link
(i.e. to an anchor starting with "#") always scrolls inside the window
to that element. Back and forth navigation has to be implemented yet.

Added little more HTML than intended yet, to make jumping to the
internal target work.